# backend/docs/swagger.yaml
basePath: /
definitions:
  api.bulkCreateRequest:
    properties:
      role:
        type: string
      users:
        items:
          $ref: '#/definitions/api.createUserRequest'
        type: array
    required:
    - users
    type: object
  api.bulkResult:
    properties:
      error:
        type: string
      password:
        type: string
      status:
        description: created | error
        type: string
      user:
        $ref: '#/definitions/user.User'
      username:
        type: string
    type: object
  api.createAdminRequest:
    properties:
      displayName:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - username
    type: object
...
  api.createTeamRequest:
    properties:
      members:
        items:
          $ref: '#/definitions/api.teamMemberRequest'
        type: array
      name:
        type: string
    required:
    - members
    - name
    type: object
  api.createUserRequest:
    properties:
      displayName:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - username
    type: object
...
  api.passwordRequest:
    properties:
      password:
        type: string
    type: object
...
  api.roleRequest:
    properties:
      role:
        type: string
    required:
    - role
    type: object
...
  api.teamMemberRequest:
    properties:
      displayName:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - username
    type: object
...
  user.User:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      role:
        type: string
      teamId:
        type: string
      teamName:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Algorithmic contest platform REST API & judge engine.
  title: MiniAlgothon API
  version: "1.0"
paths:
  /api/v1/admin/admins:
    post:
      consumes:
      - application/json
      description: Create an administrator account without a team assignment.
      parameters:
      - description: Admin account payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.createAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Create Administrator Account
      tags:
      - Admin
...
  /api/v1/admin/teams:
    post:
      consumes:
      - application/json
      description: Atomically create a team and 1 to 3 member user accounts.
      parameters:
      - description: Team creation payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.createTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Create Team with Members
      tags:
      - Admin
  /api/v1/admin/users:
    get:
      description: Fetch all user accounts and team assignments.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/user.User'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin List Users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a single user account.
      parameters:
      - description: User creation payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Create User
      tags:
      - Admin
  /api/v1/admin/users/{id}:
    delete:
      description: Delete user account.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "244":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Delete User
      tags:
      - Admin
  /api/v1/admin/users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password and invalidate all active sessions.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New password
        in: body
        name: payload
        schema:
          $ref: '#/definitions/api.passwordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Reset Password
      tags:
      - Admin
  /api/v1/admin/users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Update user role (competitor or admin).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.roleRequest'
      responses:
        "244":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin Update User Role
      tags:
      - Admin
  /api/v1/admin/users/bulk:
    post:
      consumes:
      - application/json
      description: Bulk create user accounts.
      parameters:
      - description: Bulk user creation payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/api.bulkCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/api.bulkResult'
              type: array
            type: object
      security:
      - BearerAuth: []
      summary: Admin Bulk Create Users
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"